+++ src/search/mod.rs
@@ -1,3 +1,4 @@
 pub mod ddg;
+pub mod stackexchange;
 pub mod stackoverflow;
 pub mod utils;